djangoCAS/requirements.txt
django-cas-ng==4.3.0
